requirements.txt
streamlit
openai
sounddevice
scipy
openai-whisper
python-dotenv
